+++ Vishal/Milestone 32(Graph)/02 Graph Cycle/Union_by_rank_And_Find_By_Path_Compression.cpp
@@ -1,4 +1,6 @@
-// Time Complexity - O(E*V)
+// Disjoint set
+
+// Time Complexity - O(V*log(E))
 // Time Complexity of Union and find operation - O(logE)
 
 #include<bits/stdc++.h>
@@ -11,20 +13,18 @@ class node
 
 vector<node>disjointSet;
 
+// Optimize Path Compression
 int find(int v)
 {
+    // It means it is parent of itself
     if(disjointSet[v].parent == -1)
     return v;
-    return disjointSet[v].parent = find(disjointSet[v]); // Optimization
+    return disjointSet[v].parent = find(disjointSet[v]); 
 }
 
+// Union By Rank
 int unionOP(int from, int to)
 {
-    // from = find(from);
-    // to = find(to);
-    // disjointSet[from] = to;
-
-    //Union By Rank
     //Increase rank iff rank of from and to are same.
     //Which have higher rank, will become parent.
 
@@ -48,7 +48,8 @@ bool isCyclic(vector<pair<int, int> >&edgeList)
         int fromParent = find(p.first);
         int toParent = find(p.second);
 
-        // Found a cycle
+        // If they have same absolute parent, it means they already have a path
+        // Hence we got a cycle.
         if(fromParent == toParent)
         return true;
         
@@ -72,7 +73,7 @@ int main()
         disjointSet[i].rank = 0;
     }
     
-    // Adjacent list
+    // Adjacency list
     vector<pair<int, int> >edgeList;
     for (int i = 0; i < E; i++)
     {
